Began work on server component * Renamed component: controller -> server * Integrated flask into server component * Began revising checkout process * added simulation option as "sim" and corresponding FIXME notes * Changed all indentation to tabs
