+++ main.py
@@ -1,9 +1,9 @@
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from controller import Controller
-
+import server
 
 if __name__== "__main__":
-	controller = Controller()
-	controller.setup()
-	controller.run()
+	poss = server.Server()
+	poss.setup(True)
+	poss.start(True)
