+++ server/templates/index.html.j2
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+	<head>
+		<meta charset="utf-8">
+		<title>TAPA {% block title %}{% endblock %}</title>
+	</head>
+	<body>
+
+		<p>{{ variable }}</p>
+	</body>
+</html>
