+++ tag_reader/api.py
@@ -14,111 +14,111 @@ BLOCK_SIZE = 20
 
 # PN532 is the NFC tag reader module
 class PN532(object):
-    def __init__(self):
-        # i2c device address
-        self.address = register.PN532_DEFAULT_ADDRESS
+	def __init__(self):
+		# i2c device address
+		self.address = register.PN532_DEFAULT_ADDRESS
 
-        # smbus object
-        self.bus = SMBus(1)
+		# smbus object
+		#FIXME sim
+		#self.bus = SMBus(1)
 
-        # logger object
-        self.logger = logging.getLogger()
+		# logger object
+		self.logger = logging.getLogger()
 
-    def setup(self, enable_logging=False):
-        """setup the device"""
-        # print("pn532 is setting up")
-        if enable_logging:
-            self.use_logging()
+	def setup(self, enable_logging=False):
+		"""setup the device"""
+		# print("pn532 is setting up")
+		if enable_logging:
+			self.use_logging()
 
-        self.sam_config()
+		self.sam_config()
 
-    def use_logging(self):
-        """use debug logging"""
-        # setup a custom formatting message
-        handler = logging.StreamHandler()
-        handler.setFormatter(
-            logging.Formatter("%(asctime)s | %(levelname)-5s | %(message)s")
-        )
-        self.logger.addHandler(handler)
+	def use_logging(self):
+		"""use debug logging"""
+		# setup a custom formatting message
+		handler = logging.StreamHandler()
+		handler.setFormatter(
+			logging.Formatter("%(asctime)s | %(levelname)-5s | %(message)s")
+		)
+		self.logger.addHandler(handler)
 
-        # set logging level to DEBUG
-        self.logger.setLevel(logging.DEBUG)
+		# set logging level to DEBUG
+		self.logger.setLevel(logging.DEBUG)
 
-    def read(self):
-        """return a raw reading value as an array of bytes"""
-        self.in_list_passive_target()
+	def read(self):
+		"""return a raw reading value as an array of bytes"""
+		self.in_list_passive_target()
 
-        while True:
-            read = self.read_addr(BLOCK_SIZE)
-            # check the first 3 bytes to see if a card is detected or not
-            if read[:3] != [0x00, 0x80, 0x80]:
-                # TODO - the first 9 bytes are configs bytes so we're not really
-                # interested in getting these at the moment, though they could
-                # be used for validation in the future
-                return read[9:]
+		while True:
+			read = self.read_addr(BLOCK_SIZE)
+			# check the first 3 bytes to see if a card is detected or not
+			if read[:3] != [0x00, 0x80, 0x80]:
+				# TODO - the first 9 bytes are configs bytes so we're not really
+				# interested in getting these at the moment, though they could
+				# be used for validation in the future
+				return read[9:]
 
+	def sam_config(self):
+		"""send SAMConfiguration command"""
+		self.write_addr(
+			construct_frame([register.PN532_COMMAND_SAMCONFIGURATION, 0x01, 0x01, 0x00])
+		)
+		self.read_addr(BLOCK_SIZE)
 
-    def sam_config(self):
-        """send SAMConfiguration command"""
-        self.write_addr(
-            construct_frame([register.PN532_COMMAND_SAMCONFIGURATION, 0x01, 0x01, 0x00])
-        )
-        self.read_addr(BLOCK_SIZE)
+	def in_list_passive_target(self):
+		"""send InListPassiveTarget command"""
+		self.write_addr(
+			construct_frame([register.PN532_COMMAND_INLISTPASSIVETARGET, 0x01, 0x00])
+		)
+		self.read_addr(BLOCK_SIZE)
 
-    def in_list_passive_target(self):
-        """send InListPassiveTarget command"""
-        self.write_addr(
-            construct_frame([register.PN532_COMMAND_INLISTPASSIVETARGET, 0x01, 0x00])
-        )
-        self.read_addr(BLOCK_SIZE)
+	def write_addr(self, data):
+		"""write to its own address with given block data"""
+		time.sleep(REST_INTERVAL)
 
-    def write_addr(self, data):
-        """write to its own address with given block data"""
-        time.sleep(REST_INTERVAL)
+		self.bus.write_i2c_block_data(self.address, self.address, data)
+		self.logger.debug("write_addr: %s", data)
 
-        self.bus.write_i2c_block_data(self.address, self.address, data)
-        self.logger.debug("write_addr: %s", data)
+	def read_addr(self, length):
+		"""read from its own address a given-length of block data"""
+		time.sleep(REST_INTERVAL)
 
-    def read_addr(self, length):
-        """read from its own address a given-length of block data"""
-        time.sleep(REST_INTERVAL)
+		buf = []
+		msg = i2c_msg.read(self.address, length)
+		self.bus.i2c_rdwr(msg)
 
-        buf = []
-        msg = i2c_msg.read(self.address, length)
-        self.bus.i2c_rdwr(msg)
+		for b in msg:
+			buf.append(b)
 
-        for b in msg:
-            buf.append(b)
+		self.logger.debug("read_addr: %s", buf)
+		return buf
 
-        self.logger.debug("read_addr: %s", buf)
-        return buf
+	def sim_setup(self):
+		"""FIXME - simulate setup, only used for testing purposes"""
+		pass
 
-    def simulate_setup(self):
-        """FIXME - simulate setup, only used for testing purposes"""
-        return [100, 0, 0]
-
-    def simulate_read(self):
-        """FIXME - simulate reading, only used for testing purposes"""
-        return [100, 0, 0]
-
+	def sim_read(self):
+		"""FIXME - simulate reading, only used for testing purposes"""
+		time.sleep(Rest)
+		return {1,0,68,0,7,4,137,16,98,101,96}
 
 def construct_frame(data):
-    """construct frame for communicating between host controller and pn532"""
-    # begin with 6-bytes frame structure
-    buf = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00]
-    buf[0] = register.PN532_PREAMBLE
-    buf[1] = register.PN532_STARTCODE1
-    buf[2] = register.PN532_STARTCODE2
-    buf[3] = len(data) + 1           # number of bytes in data and frame identifier field
-    buf[4] = (~buf[3] & 0xFF) + 0x01 # packet length checksum
-    buf[5] = register.PN532_HOSTTOPN532
+	"""construct frame for communicating between host controller and pn532"""
+	# begin with 6-bytes frame structure
+	buf = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00]
+	buf[0] = register.PN532_PREAMBLE
+	buf[1] = register.PN532_STARTCODE1
+	buf[2] = register.PN532_STARTCODE2
+	buf[3] = len(data) + 1		   # number of bytes in data and frame identifier field
+	buf[4] = (~buf[3] & 0xFF) + 0x01 # packet length checksum
+	buf[5] = register.PN532_HOSTTOPN532
 
-    tmp_sum = register.PN532_HOSTTOPN532
-    for b in data:
-        tmp_sum += b
-        buf.append(b)
+	tmp_sum = register.PN532_HOSTTOPN532
+	for b in data:
+		tmp_sum += b
+		buf.append(b)
 
-    buf.append((~tmp_sum & 0xFF) + 0x01) # data checksum
-    buf.append(register.PN532_POSTAMBLE)
+	buf.append((~tmp_sum & 0xFF) + 0x01) # data checksum
+	buf.append(register.PN532_POSTAMBLE)
 
-    return buf
+	return buf
